WIP: Attempts to run both live & simulated robots
Tried to run gazebo in a namespace and connect the published robot states and controllers via remaps. Did not work, cannot find the source of the problems. MoveIt complains about the planning scene missing (although the package did not change(!)) and the real robot controller might not even be connected yet.